+++ http-client/test-nonet/Network/HTTP/Client/RequestSpec.hs
@@ -99,14 +99,14 @@ spec = do
                , [nonChunked <> "\r\ndata"]
                , [nonChunked <> "\r\ndata"]
                , [nonChunked <> "\r\n", "", "data"]
-               , [chunked <> "\r\n", "", "4\r\n","data","\r\n","0\r\n\r\n"]
+               , [chunked <> "\r\n", "", "4\r\ndata\r\n","0\r\n\r\n"]
                ]
 
         out2 = [ (nonChunked <> "Expect: 100-continue\r\n\r\n", ["data"])
                , (nonChunked <> "Expect: 100-continue\r\n\r\n", ["data"])
                , (nonChunked <> "Expect: 100-continue\r\n\r\n", ["data"])
                , (nonChunked <> "Expect: 100-continue\r\n\r\n", ["data"])
-               , (chunked    <> "Expect: 100-continue\r\n\r\n", ["4\r\n","data","\r\n","0\r\n\r\n"])
+               , (chunked    <> "Expect: 100-continue\r\n\r\n", ["4\r\ndata\r\n","0\r\n\r\n"])
                ]
 
         nonChunked = "PUT /foo HTTP/1.1\r\nHost: localhost\r\nAccept-Encoding: gzip\r\nContent-Length: 4\r\n"